refactor: factor common sdam types into commonly accesible file
