+++ lib/core/sdam/common.js
@@ -1,0 +1,38 @@
+'use strict';
+
+// shared state names
+const STATE_CLOSING = 'closing';
+const STATE_CLOSED = 'closed';
+const STATE_CONNECTING = 'connecting';
+const STATE_CONNECTED = 'connected';
+
+// An enumeration of topology types we know about
+const TopologyType = {
+  Single: 'Single',
+  ReplicaSetNoPrimary: 'ReplicaSetNoPrimary',
+  ReplicaSetWithPrimary: 'ReplicaSetWithPrimary',
+  Sharded: 'Sharded',
+  Unknown: 'Unknown'
+};
+
+// An enumeration of server types we know about
+const ServerType = {
+  Standalone: 'Standalone',
+  Mongos: 'Mongos',
+  PossiblePrimary: 'PossiblePrimary',
+  RSPrimary: 'RSPrimary',
+  RSSecondary: 'RSSecondary',
+  RSArbiter: 'RSArbiter',
+  RSOther: 'RSOther',
+  RSGhost: 'RSGhost',
+  Unknown: 'Unknown'
+};
+
+module.exports = {
+  STATE_CLOSING,
+  STATE_CLOSED,
+  STATE_CONNECTING,
+  STATE_CONNECTED,
+  TopologyType,
+  ServerType
+};
