+++ lib/core/sdam/monitoring.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const ServerDescription = require('./server_description').ServerDescription;
+const ServerDescription = require('./server_description');
 const calculateDurationInMs = require('../utils').calculateDurationInMs;
 
 // pulled from `Server` implementation
