+++ lib/core/sdam/server.js
@@ -7,7 +7,7 @@ const wireProtocol = require('../wireprotocol');
 const BSON = require('../connection/utils').retrieveBSON();
 const createClientInfo = require('../topologies/shared').createClientInfo;
 const Logger = require('../connection/logger');
-const ServerDescription = require('./server_description').ServerDescription;
+const ServerDescription = require('./server_description');
 const ReadPreference = require('../topologies/read_preference');
 const monitorServer = require('./monitoring').monitorServer;
 const MongoParseError = require('../error').MongoParseError;
@@ -16,6 +16,7 @@ const collationNotSupported = require('../utils').collationNotSupported;
 const debugOptions = require('../connection/utils').debugOptions;
 const isSDAMUnrecoverableError = require('../error').isSDAMUnrecoverableError;
 const makeStateMachine = require('../utils').makeStateMachine;
+const common = require('./common');
 
 // Used for filtering out fields for logging
 const DEBUG_FIELDS = [
@@ -45,10 +46,10 @@ const DEBUG_FIELDS = [
   'servername'
 ];
 
-const STATE_CLOSING = 'closing';
-const STATE_CLOSED = 'closed';
-const STATE_CONNECTING = 'connecting';
-const STATE_CONNECTED = 'connected';
+const STATE_CLOSING = common.STATE_CLOSING;
+const STATE_CLOSED = common.STATE_CLOSED;
+const STATE_CONNECTING = common.STATE_CONNECTING;
+const STATE_CONNECTED = common.STATE_CONNECTED;
 const stateTransition = makeStateMachine({
   [STATE_CLOSED]: [STATE_CLOSED, STATE_CONNECTING],
   [STATE_CONNECTING]: [STATE_CONNECTING, STATE_CLOSING, STATE_CONNECTED, STATE_CLOSED],
