+++ lib/core/sdam/server_description.js
@@ -2,19 +2,7 @@
 
 const arrayStrictEqual = require('../utils').arrayStrictEqual;
 const tagsStrictEqual = require('../utils').tagsStrictEqual;
-
-// An enumeration of server types we know about
-const ServerType = {
-  Standalone: 'Standalone',
-  Mongos: 'Mongos',
-  PossiblePrimary: 'PossiblePrimary',
-  RSPrimary: 'RSPrimary',
-  RSSecondary: 'RSSecondary',
-  RSArbiter: 'RSArbiter',
-  RSOther: 'RSOther',
-  RSGhost: 'RSGhost',
-  Unknown: 'Unknown'
-};
+const ServerType = require('./common').ServerType;
 
 const WRITABLE_SERVER_TYPES = new Set([
   ServerType.RSPrimary,
@@ -186,7 +174,4 @@ function parseServerType(ismaster) {
   return ServerType.Standalone;
 }
 
-module.exports = {
-  ServerDescription,
-  ServerType
-};
+module.exports = ServerDescription;
