+++ lib/core/sdam/server_selectors.js
@@ -1,6 +1,6 @@
 'use strict';
-const ServerType = require('./server_description').ServerType;
-const TopologyType = require('./topology_description').TopologyType;
+const ServerType = require('./common').ServerType;
+const TopologyType = require('./common').TopologyType;
 const ReadPreference = require('../topologies/read_preference');
 const MongoError = require('../error').MongoError;
