+++ lib/core/sdam/topology.js
@@ -1,9 +1,9 @@
 'use strict';
 const EventEmitter = require('events');
-const ServerDescription = require('./server_description').ServerDescription;
-const ServerType = require('./server_description').ServerType;
-const TopologyDescription = require('./topology_description').TopologyDescription;
-const TopologyType = require('./topology_description').TopologyType;
+const ServerDescription = require('./server_description');
+const ServerType = require('./common').ServerType;
+const TopologyDescription = require('./topology_description');
+const TopologyType = require('./common').TopologyType;
 const monitoring = require('./monitoring');
 const calculateDurationInMs = require('../utils').calculateDurationInMs;
 const MongoTimeoutError = require('../error').MongoTimeoutError;
@@ -27,6 +27,7 @@ const SrvPoller = require('./srv_polling').SrvPoller;
 const getMMAPError = require('../topologies/shared').getMMAPError;
 const makeStateMachine = require('../utils').makeStateMachine;
 const eachAsync = require('../utils').eachAsync;
+const common = require('./common');
 
 // Global state
 let globalTopologyCounter = 0;
@@ -62,10 +63,10 @@ const LOCAL_SERVER_EVENTS = SERVER_RELAY_EVENTS.concat([
   'ended'
 ]);
 
-const STATE_CLOSING = 'closing';
-const STATE_CLOSED = 'closed';
-const STATE_CONNECTING = 'connecting';
-const STATE_CONNECTED = 'connected';
+const STATE_CLOSING = common.STATE_CLOSING;
+const STATE_CLOSED = common.STATE_CLOSED;
+const STATE_CONNECTING = common.STATE_CONNECTING;
+const STATE_CONNECTED = common.STATE_CONNECTED;
 const stateTransition = makeStateMachine({
   [STATE_CLOSED]: [STATE_CLOSED, STATE_CONNECTING],
   [STATE_CONNECTING]: [STATE_CONNECTING, STATE_CLOSING, STATE_CONNECTED, STATE_CLOSED],
