+++ lib/core/sdam/topology_description.js
@@ -1,22 +1,14 @@
 'use strict';
-const ServerType = require('./server_description').ServerType;
-const ServerDescription = require('./server_description').ServerDescription;
+const ServerType = require('./common').ServerType;
+const ServerDescription = require('./server_description');
 const WIRE_CONSTANTS = require('../wireprotocol/constants');
+const TopologyType = require('./common').TopologyType;
 
 // contstants related to compatability checks
 const MIN_SUPPORTED_SERVER_VERSION = WIRE_CONSTANTS.MIN_SUPPORTED_SERVER_VERSION;
 const MAX_SUPPORTED_SERVER_VERSION = WIRE_CONSTANTS.MAX_SUPPORTED_SERVER_VERSION;
 const MIN_SUPPORTED_WIRE_VERSION = WIRE_CONSTANTS.MIN_SUPPORTED_WIRE_VERSION;
 const MAX_SUPPORTED_WIRE_VERSION = WIRE_CONSTANTS.MAX_SUPPORTED_WIRE_VERSION;
-
-// An enumeration of topology types we know about
-const TopologyType = {
-  Single: 'Single',
-  ReplicaSetNoPrimary: 'ReplicaSetNoPrimary',
-  ReplicaSetWithPrimary: 'ReplicaSetWithPrimary',
-  Sharded: 'Sharded',
-  Unknown: 'Unknown'
-};
 
 // Representation of a deployment of servers
 class TopologyDescription {
@@ -404,7 +396,4 @@ function checkHasPrimary(serverDescriptions) {
   return TopologyType.ReplicaSetNoPrimary;
 }
 
-module.exports = {
-  TopologyType,
-  TopologyDescription
-};
+module.exports = TopologyDescription;
