+++ lib/core/topologies/shared.js
@@ -3,7 +3,7 @@
 const os = require('os');
 const ReadPreference = require('./read_preference');
 const Buffer = require('safe-buffer').Buffer;
-const TopologyType = require('../sdam/topology_description').TopologyType;
+const TopologyType = require('../sdam/common').TopologyType;
 const MongoError = require('../error').MongoError;
 
 const MMAPv1_RETRY_WRITES_ERROR_CODE = 20;
