+++ lib/core/wireprotocol/shared.js
@@ -2,8 +2,8 @@
 
 const ReadPreference = require('../topologies/read_preference');
 const MongoError = require('../error').MongoError;
-const ServerType = require('../sdam/server_description').ServerType;
-const TopologyDescription = require('../sdam/topology_description').TopologyDescription;
+const ServerType = require('../sdam/common').ServerType;
+const TopologyDescription = require('../sdam/topology_description');
 
 const MESSAGE_HEADER_SIZE = 16;
 const COMPRESSION_DETAILS_SIZE = 9; // originalOpcode + uncompressedSize, compressorID
