+++ test/core/unit/sdam_spec_tests.js
@@ -5,7 +5,7 @@ const chai = require('chai');
 const expect = chai.expect;
 const Topology = require('../../../lib/core/sdam/topology');
 const Server = require('../../../lib/core/sdam/server');
-const ServerDescription = require('../../../lib/core/sdam/server_description').ServerDescription;
+const ServerDescription = require('../../../lib/core/sdam/server_description');
 const monitoring = require('../../../lib/core/sdam/monitoring');
 const parse = require('../../../lib/core/uri_parser');
 chai.use(require('chai-subset'));
